+++ src/components/layout/index.tsx
@@ -1,0 +1,2 @@
+export { default as AppLayout } from './AppLayout/AppLayout';
+export { default as PurchaseProcessLayout } from './PurchaseProcessLayout/PurchaseProcessLayout';
