+++ src/components/orderConfirm/CouponModal/CouponModal.tsx
@@ -2,12 +2,12 @@ import { Coupon } from '@appTypes/shoppingCart';
 import { CrossIcon } from '@assets/index';
 import { InfoBanner } from '@components/common';
 import { MAX_NUMBER_OF_COUPON } from '@constants/coupon';
-import { useMaxDiscountCalculator, useModalTargetEl, usePreventScroll } from '@hooks/index';
-import { couponListAtom, maxDiscountAtom, selectedCouponsAtom } from '@recoil/shoppingCart';
+import { useCouponFinder, useMaxDiscountCalculator, useModalTargetEl, usePreventScroll } from '@hooks/index';
+import { maxDiscountAtom, selectedCouponsAtom } from '@recoil/shoppingCart';
 import { formatKoreanCurrency } from '@utils/currency';
 import { CenterModal, ModalContainer } from 'badahertz52-react-modules-components';
-import { ChangeEvent, Dispatch, SetStateAction, useEffect, useState } from 'react';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { ChangeEvent, Dispatch, SetStateAction, useState } from 'react';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 
 import CouponCard from '../CouponCard/CouponCard';
 
@@ -18,13 +18,15 @@ interface CouponModalProps {
   setOpenModal: Dispatch<SetStateAction<boolean>>;
   availableCoupons: Coupon[];
 }
+
 const CouponModal = ({ openModal, setOpenModal, availableCoupons }: CouponModalProps) => {
-  const couponMap = useRecoilValue(couponListAtom);
   const setMaxDiscount = useSetRecoilState(maxDiscountAtom);
   const [selectedCoupons, setSelectedCoupons] = useRecoilState(selectedCouponsAtom);
 
   const { getMaxDiscountAmount } = useMaxDiscountCalculator();
   const { modalTargetEl } = useModalTargetEl();
+  const { getCoupon, getAllCoupons } = useCouponFinder();
+
   usePreventScroll({ targetEl: document.body as HTMLBodyElement, isPreventScroll: openModal });
 
   const [selectedCouponCodes, setSelectedCouponCodes] = useState<string[]>(
@@ -32,7 +34,7 @@ const CouponModal = ({ openModal, setOpenModal, availableCoupons }: CouponModalP
   );
   const [discount, setDiscount] = useState<number>(0);
 
-  const coupons = Array.from(couponMap.values());
+  const coupons = Object.values(getAllCoupons());
   const availableCouponCodes = availableCoupons.map((coupon) => coupon.code);
 
   const isCouponDisabled = (code: string) => {
@@ -40,27 +42,47 @@ const CouponModal = ({ openModal, setOpenModal, availableCoupons }: CouponModalP
     if (!availableCouponCodes.includes(code)) return true;
     // 이미 최대 사용 가능한 쿠폰 개수만큼 쿠폰을 선택했지만, 해당 쿠폰은 선택되지 않은 경우
     if (selectedCouponCodes.length >= MAX_NUMBER_OF_COUPON && !selectedCouponCodes.includes(code)) return true;
+
     return false;
   };
 
-  const handleChangeChecked = (e: ChangeEvent<HTMLInputElement>, code: string) => {
-    const isChecked = e.target.checked;
+  /**
+   * 쿠폰 선택 변경에 따른 변경될 selectedCouponCodes를 반환하는 함수
+   * @param isChecked
+   * @param code
+   */
+  const getNewSelectedCouponCodes = (isChecked: boolean, code: string) => {
+    if (isChecked) return selectedCouponCodes?.includes(code) ? selectedCouponCodes : selectedCouponCodes.concat(code);
 
-    setSelectedCouponCodes((prev) => {
-      if (isChecked) return prev?.includes(code) ? prev : prev.concat(code);
-
-      return prev.filter((i) => i !== code);
-    });
+    return selectedCouponCodes.filter((i) => i !== code);
   };
 
-  const updateMaxDiscount = () => {
-    const selectedCoupons = selectedCouponCodes
-      .map((code) => couponMap.get(code))
+  /**
+   * 선택된 쿠폰에 따른 최대 할인 금액을 계산해, 로컬 상태인 discount를 업데이트하는 함수
+   */
+  const updateDiscount = (newSelectedCouponCodes: string[]) => {
+    const selectedCoupons = newSelectedCouponCodes
+      .map((code) => getCoupon(code))
       .filter((coupon): coupon is Coupon => coupon !== undefined);
 
     setDiscount(getMaxDiscountAmount(selectedCoupons));
   };
 
+  /**
+   * 쿠폰에 대한 체크 박스의 선택 여부를 변경 시, 실행되는 기능(selectedCouponCodes,discount)에 대한 핸들러
+   * @param code :해당 체크 박스의 타켓인 쿠폰의 코드
+   */
+  const handleChangeChecked = (e: ChangeEvent<HTMLInputElement>, code: string) => {
+    const isChecked = e.target.checked;
+
+    const newSelectedCouponCodes = getNewSelectedCouponCodes(isChecked, code);
+    setSelectedCouponCodes(newSelectedCouponCodes);
+    updateDiscount(newSelectedCouponCodes);
+  };
+
+  /**
+   * 쿠폰 모달에서 선택된 쿠폰을 바탕으로 selectedCoupons 상태를 업데이트하는 함수
+   */
   const updateSelectedCoupons = () => {
     const selectedCoupons = availableCoupons.filter((coupon) => selectedCouponCodes.includes(coupon.code));
     setSelectedCoupons(selectedCoupons);
@@ -74,10 +96,6 @@ const CouponModal = ({ openModal, setOpenModal, availableCoupons }: CouponModalP
     // 모달 닫기
     setOpenModal(false);
   };
-
-  useEffect(() => {
-    updateMaxDiscount();
-  }, [selectedCouponCodes]);
 
   return (
     <>
@@ -104,7 +122,7 @@ const CouponModal = ({ openModal, setOpenModal, availableCoupons }: CouponModalP
             ))}
           </Styled.CouponList>
           <Styled.ApplyCouponButton onClick={handleClickApplyCoupons}>
-            총 {formatKoreanCurrency(discount)}할인 쿠폰 사용하기
+            총 {formatKoreanCurrency(discount)} 할인 쿠폰 사용하기
           </Styled.ApplyCouponButton>
         </CenterModal>
       )}
