+++ src/components/shoppingCart/index.ts
@@ -1,6 +1,5 @@
-export { default as CountAlertModal } from './CountAlertModal/CountAlertModal';
 export { default as AmountsList } from './AmountsList/AmountsList';
-export { default as PurchaseProcessLayout } from '@components/layout/PurchaseProcessLayout/PurchaseProcessLayout';
+export { default as CountAlertModal } from './CountAlertModal/CountAlertModal';
 export { default as CartListDescription } from './CartListDescription/CartListDescription';
 export { default as CartList } from './CartList/CartList';
 export { default as CartListItem } from './CartListItem/CartListItem';
