+++ src/hooks/coupon/useCouponDiscount.ts
@@ -4,10 +4,6 @@ import { PRICE } from '@constants/shippingCart';
 import useBuyXgetYTargetItem from './useBuyXgetYTargetItem';
 import useCouponFinder from './useCouponFinder';
 
-/**
- * 쿠폰 적용 대상 금액에 대해 쿠폰 적용 시 할인된 금액을 반환하는 훅
- * @param param0
- */
 const useCouponDiscount = () => {
   const { getCoupon } = useCouponFinder();
   const { getBuyXgetYTargetItem } = useBuyXgetYTargetItem();
@@ -38,16 +34,25 @@ const useCouponDiscount = () => {
     return targetItem.product.price;
   };
 
+  /**
+   * 쿠폰 적용 대상 금액에 대해 쿠폰 적용에 따른 할인 금액을 반환하는 훅
+   * @param couponCode : 적용되는 쿠폰의 코드
+   * @param couponTargetPrice  : 쿠폰 적용 대상인 금액
+   * @returns : 할인 금액
+   */
   const calculateCouponDiscount = (couponCode: CouponCode, couponTargetPrice?: number) => {
     const coupon = getCoupon(couponCode);
 
     switch (coupon.discountType) {
       case 'fixed':
         return calculateFixedDiscount(coupon);
+
       case 'buyXgetY':
         return calculateBuyXgetYDiscount();
+
       case 'freeShipping':
         return PRICE.shippingFee.basic;
+
       case 'percentage':
         return calculatePercentageDiscount(coupon, couponTargetPrice ?? 0);
