+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
-export { default as useFetchErrorBoundary } from './useFetchErrorBoundary';
 export * from './coupon';
 export * from './modal';
 export * from './shoppingCart';
+export { default as useFetchErrorBoundary } from './useFetchErrorBoundary';
