+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export { default as ErrorPage } from './ErrorPage/ErrorPage';
-export { default as PurchaseConfirmPage } from './PurchaseConfirmPage/PurchaseConfirmPage';
 export { default as OrderConfirmPage } from './OrderConfirmPage/OrderConfirmPage';
 export { default as OrderPage } from './OrderPage/OrderPage';
+export { default as PurchaseConfirmPage } from './PurchaseConfirmPage/PurchaseConfirmPage';
